Fixed env build scripts.
Many of the scripts would fall off the end of a loop without finding anything but then report a success code.  The BOOST_PYTHON script was having problems with this, leaving the BOOST_PYTHON variable silently defined to the empty string.  Added exit 1 to a lot of scripts.  Also fixed the problem causing BOOST_PYTHON to fail in the first place.