Removed (size_t) casts done for Boost 1.35.
Boost's iterator_range::size() returns difference_type instead of
size_type in Boost 1.35.  I guess that makes some sense in the context
of a general iterator_range, but for my use as a substring range, it's
annoying.  However, I'm no longer using Boost for that.  There's a new
class, StringRange.  So I removed the hacks to make to code work with
Boost.