Delayed tablet config updates until next sync() call.
Tablet configs used to be saved immediately.  Now they're delayed
until the next call to sync().  This seems to fix the deadlock
problem, but it opens up some other issues.  For example, what happens
if the server drops a tablet before the client calls sync(), but then
the client subsequently calls sync().  I think the server will reload
the tablet with the old config, potentially losing mutations in the
pipe, and then tell the client the sync succeeded.  This is bad.  The
server shouldn't drop tablets with changes pending, or sync them
before doing so.