Fixed potential deadlock problem with SyncQueue::waitForCompletion()
waitForCompletion() was waiting on the taskCompleted condition
variable.  This condition would be signaled once for every
taskComplete() call.  If multiple threads are in waitForCompletion(),
only one will wake up for each taskComplete().  Usually it'll check to
see if the nTasksPending is zero or go back to sleep.

I changed the taskComplete condition to allTasksComplete, which
notifies all waiting threads once nTasksPending hits zero.