Simplied loading of meta table in MetaConfigManager.
The old method needed the META table to be specified as one of several
special fixed tables.  Furthermore, all servers would connect to the
META table through the same NetTable interface.  For the server
hosting the META table, this introduced problems.  It would need to
call out to itself to access the META table, which could block
indefinitely if the thread pool was full of operations waiting on META
table.

The new scheme gives the MetaConfigManager a direct TablePtr to the
META table.  For the server hosting the META table, this will be a
TabletPtr.  For remote servers, it will be a NetTable.  The
SuperTabletMaker knows how to return a network-accessible version of
the META table for remote servers that need it.

This change simplifies the design a bit and fixes some deadlock
problems.  There are still other deadlock problems with the serialize
loop calling replaceTables(), which in turn calls saveConfig().