Fixed problem where multiple Tables could be created for the same ID
Multiple objects could exist simultaneously for the same ID.  The old
implementation would at least discard the objects that turned out to
be duplicates, but it still wasn't safe.  Tablets can rewrite their
config on load, so if the same Tablet is loaded 4 times simultaneously
and only one of them is kept, that leads to badness.  Changed the
implementation to wait for other threads currently working on creating
the same object.