WIP: This code actually performed a tablet split!
First dynamic partition!  Yay!  Also, changed the kdi: scheme to use
"localhost" as the default host and "10000" as the default port.

Issues: code still doesn't use SuperScanner.  After the partition, the
client usually barfs on a scan because the scan prediate maps to
multiple tables.  Also, it's possible for the META table to get out of
whack if Tablets get destroyed without saving pending config changes.