Tablet config changes are now saved ASAP.
Replaced SharedSplitter with WorkQueue, which handles splits as well
as deferred saveConfig() calls.  This ensures that tablet config
changes are saved in a timely manner.  Also, it shouldn't be possible
for a Tablet to get deleted with a pending config change, since the
deferred save config call will hold a shared pointer to the Tablet.