First pass KDI-Hypertable bridge.
KDI has a more permissive data API than Hypertable, so the adapter
needs to do some encoding to store things in Hypertable.  It seems
like there are some bugs at the moment.  For example, setting,
erasing, and setting the same cells doesn't seem to work.  The cells
get lost.  I think this may be a HT problem, but not sure.