Added work-around for Hypertable timeouts for slow scanners.
KDI will reopen the Hypertable scan if it finds that the scanner has
timed out when it tries to get the next cell.  If the scan has to be
re-opened, KDI will advance the scan so that the next cell returned
will be after the last cell previously returned to the client.