Made library and application targets require object inputs.
If you typed "make foo", make would assume it could build an
application "foo" and try to link it with no objects.  It would fail,
but it was a weird error message.  Typing "make libfoo.a" or "make
libfoo.so" would succeed in creating a bogus empty library in your
source directory.  Now all of these cases fail with a descriptive
error message.