Bug fix: tracker was prematurely deleting fragment files
There was a problem with SharedLogger where it would add a newly
serialized file to a tablet before registering it with FileTracker.
If the tablet happened to get split between adding the serialized
fragment and tracking it, the newly split tablet would clone and
reference the fragment.  The additional reference would be ignored
since the file was not tracked.  Then when one of the tablets released
the file, FileTracker would delete the file, even though the other
tablet still needed it.  This problem would sometimes show up as an
error when the compactor tried to read the fragment that had been
deleted.  The simple fix was to track the file before adding the
fragment to a tablet.

Also added more logging on fragment shuffling and file tracking.

There is still an outstanding bug with split tablets.  The split
tablet will clone the base tablet's log fragment, if any.  However,
the SharedLogger won't know about the split, so when it comes time to
serialize the fragment, the cloned tablet won't have that fragment
replaced.

This bug breaks memory usage assumptions and can lead to data loss.
The memory problem is that tablets wind up holding references to
in-memory cells for longer than we expect.  The tablets will hold the
fragments until the compactor comes around to the tablet and replaces
the log fragment.  Also, if no other data is ever added to the tablet,
the compactor may never actually replace the log fragment.

Data loss and state corruption can occur if the server exits before
all the dangling log fragments can be compacted.  The log file is
deleted immediately after the data is serialized, but some tablets
still hold the fragment in memory.  If the process exits, this data is
lost.  Also, there will be problems reloading the tablet since its
state will be referencing log files that have been deleted.

Fixing this problem will take some SharedLogger restructuring.