Fixed problems with cloning log fragments in splits
There was some badness that happened when log fragments got cloned as
part of a Tablet split.  See commit notes from b9404337dd.

These problems have been resolved by tracking when a log fragment is
cloned.  Replacement notifications are forwarded to the split-off
Tablet that cloned the fragment so it is sure to get the update too.
In addition, Tablets are now able to detect duplicate adds of log
fragments.  This can happen an update arrives in the same log after
the Tablet has been split.  In this case, we can be sure that the
SharedLogger knows about the Tablet and will provide its own
replacement notification, so there's an extra flag to ignore the
chained update from the origin Tablet.

This is all a bit tricky, but it seems to resolve all the problems.
It's ugly, though, because there are some unspoken agreements on how
Tablet and SharedLogger have to behave.  For example, SharedLogger
must give the same LogFragment to every Tablet in the same table, or
things will break.