Made kdiNetServer (SuperTablet) create tables that don't exist
The old behavior was to throw an error about the table not exising,
though that error never got propagated to the client in any infomative
way.  Now if the table doesn't exist, the server will insert it into
the META table automatically.  This is more consistent with normal KDI
usage.