Added HashedTable (KDI "hash" scheme)
HashedTable is a client-side adapter that distributes a logical table
over a fixed set of servers by hashing row keys.  The server set
cannot change or things will get broken.  Scans have to read from all
servers and merge at the client.  All in all, it's not an ideal way to
do table distribution, but it's dead simple and it works.