Stop normalizing URIs passed through 'sync' scheme
The open function for KDI SynchronizedTable uses a cache to map open()
requests for the same URI to same underlying table.  The URI is
normalized before checking in the cache, which makes some sense.
However, the normalization doesn't play well with all sub-schemes.
One normalization is to lowercase the host section.  This causes
problems with the 'hash' scheme when using a capital variable in the
host segment.  The fix was to normalize the cache key, but use the
original URI for opening the sub-table.