Fixed warp::fs::replacePath() bug
replacePath() would replace the entire path section of the URI with
whatever string you gave it.  For URIs with an authority segment,
relative paths would get joined directly to the authority.  That is,
replacePath("//x/a","b") would yield "//xb".  Fixed this to produce
"//x/b".  The presence of a URI authority segment implies that the
path is rooted.  Calling replacePath() on any URI with an authority
will result in an absolute path, even if the given path is relative or
empty.