Added scanIntervals() to Table API, added RowInterval class.
Table::scanIntervals() can be used to get decent row partitions for
scheduling a distributed job.  This commit adds the basic API along
with a default, single all-encompassing interval implementation.

The result of scanIntervals() is a stream of RowInterval objects.  A
RowInterval is a simple extension of warp::Interval<std::string> that
supports serialization as a row predicate string.

Also added a command line tool, kdiIntervals, which scans and prints
intervals for tables.