Fixed file ref-counting problem in tablet server.
There was a case where a file ref could get lost, causing a file to
get deleted while it was still in use.  It happened when a Tablet
cloned a log file but didn't get a later mutation in the same log
before serialization.  In that case, the originating tablet would
forward the replaceFragments() call from SharedLogger, but not add an
additional reference.

Simplified this a bit: Tablets *always* add a reference on
addFragment() and replaceFragments().  It used to be outside code that
did this.  Now the outside code initiates the ref-count with the call
to track, but releases the file as soon as it has given everyone an
opportunity to add their own reference.