Fixed race where scanners read deleted files
In Tablet::replaceFragments(), old fragments are marked for deletion
and scanners are told to restart on the new fragment set.  Once the
fragments are marked, another thread will do the actual deletion.  If
that thread wakes up between marking and updating the scanners, the
scanners can try to read missing files.  Fixed that by postponing the
mark-for-release process until the scanners have been updated.