Wrote skeleton for Tablet garbage collection tool
The tool scans tablet server's data directory and looks for old files
that are no longer used.  A file is still in use if it is referenced
by some tablet config.

This initial checkin has most of the logic in place, but it has never
been compiled and obviously needs some massaging (missing include
files, namespaces, etc).