Made kdiTabletGc delete stuff; fixed some path bugs.
Scattered some debug printing around to make sure the paths were
right.  Tried a few test cases, flushed out obvious bugs, and added
some teeth to the file deletion message in the form of a call to
fs::remove().  This has been run on a few active table servers and
seems to do the right thing.