kdiNetServer: avoid repeated replay for shared log fragments
Added CachedLogLoader, which will load a "sharedlog:" fragment URI,
serialize it to a "disk:" fragment, and cache and return the result.

Replaced the log replay logic in ConfigManager to use CachedLogLoader.
The net effect is that each shared log will be replayed once per table
instead of once for each tablet.