Refactored MetaConfigManager to avoid smart pointer cycles
In the context of kdiNetServer, the MetaConfigManager would be
created, a FixedAdapter created from that, and then a Tablet for the
META table would use the adapter.  Then MetaConfigManager would be
given a handle to the META table.  So the META table had a handle to
the FixedAdapter, which had a handle to the MetaConfigManager, which
had a handle to the META table.  Oops.  This would prevent things from
getting cleaned up properly on shutdown.  It was a one-time leak.
Fixed that by removing FixedAdapter's handle to the MetaConfigManager.
It doesn't actually need it.

There was another loop: SharedLogger had a handle to a ConfigManager,
which, for the MetaConfigManager, has handle to the META table, which
has a handle to the SharedLogger.  This one is less easy to break.
Instead, I made SharedLogger release its handle on shutdown().

This server code has grown into a complicated mess of entangled
objects.  Now that I have some idea of what the server should be
doing, I should probably try to start with a fresh design.  Someday.

Also added/updated some log messages.