Generalized Builder offsets; other minor Builder updates.
An offset can now refer to a position within the target block instead
of just the head of the block.

Added new private function: BuilderBlock::grow(), which used to be
part of write().

Added BuilderBlock::appendPadding() which grows the block to fill out
an alignment requirement.

Added documentation to BuilderBlock private methods.

Added Builder unit tests.