Reduced allocations in DynamicCell by half
Each Cell through KDI is usually a DynamicCell.  When creating a
DynamicCell, there were two allocations: first the DynamicCell class
was created, then the buffer to hold the cell data.  Now these are
rolled into one.  The DynamicCell object is written into the header of
a buffer shared with the cell data.