Added Cell field proxies for use with STL algos
The Cell field proxy is an object that can be compared against a Cell.
When the comparison is performed, it is restricted to the field in
question.  For example:

   CellColumn("foo") == cell  <==>  StringRange("foo") == cell.getColumn()

This is useful with STL algorithms that take a comparable objects as a
parameter.  Suppose you have a vectoer of Cells and you want to find
the first Cell with a certain value:

   iter = std::find(v.begin(), v.end(), CellValue("val"));

The following Cell field proxies are defined: CellRow, CellColumn,
CellColumnFamily, CellColumnQualifier, CellTimestamp, and CellValue.

Also created prefix field proxies for the string fields, which only
match up to the specified prefix.  So this would find the first cell
with the given column prefix:

   iter = std::find(v.begin(), v.end(), CellColumnPrefix("anchor:"));

(of course, in the case, one could have also used CellColumnFamily).