Fixed Timestamp timezone handling as much as possible
The way I used to handle time zones is to mktime() using whatever the
local timezone is, figure out the difference between the input
timezone and the local timezone, then offset the timestamp.  Some
conversions for dates before 1883 were getting broken in glibc 2.8
that used to work in 2.4.  I'm not sure what the significance of
1882/1883 is.

Digging around a bit, it appears that the portable way to get a
timestamp in another timezone is to set the TZ environment variable to
a string describing the input timezone before calling mktime().  Hmm,
that's great and all, but it tosses thread-safety out the window.

That annoyed me until I realized that time handling in glibc (and more
generally, as defined by POSIX) is a mess of races on global state.
At that point, annoyance gave way to resignation.

So here's what I know: mktime() implicitly calls tzset(), which reads
the environment variable "TZ", looks at some files, and figures out
timezone information.  So if you want to use mktime(), you have to
cross your fingers that no one else happens to be mucking with TZ at
the same time.  Glibc does some locking, so things won't crash if
someone messes with TZ or calls a function that implies tzset() at
same time.  But there's a still a race: there's no guarantee that the
timezone you just set is going to be the timezone used by mktime().

The functions localtime() and localtime_r() also use the timezone
global data.  localtime() has the same issue as mktime().  It
implicitly calls tzset() for you, but you still have the race.  It
also has the obvious data-corruption race of returning a pointer to a
static structure.  localtime_r() avoids the static structure problem,
but doesn't help with the timezone issue.  In fact, it's actually a
bit worse.  The standard says that localtime_r() doesn't have to call
tzset(), but MAY.  Some message thread suggested that the wording on
the standard has since been upgraded to SHOULD.  In glibc 2.8,
localtime_r() will call tzset() if something else hasn't already, but
then it won't call it again.  So that means localtime_r() will use
whatever temporary timezone someone happened to throw at another call
to mktime(), localtime(), or tzset(), regardless of the current TZ
setting.  Defensive code should always call tzset() before
localtime_r() -- it doesn't avoid the timezone race, but it at least
restricts the window.

Applications wanting to do thread-safe time handling in multiple
timezones will need to implement higher-level locking over the libc
time functions.  warp::Timestamp tries to use these libc time
functions as safely as possible, but it's definitely not thread safe.