Fixed deadlock in Tablet::splitTablet()
Oddly enough, the deadlock was caused on a line commented with "XXX
should unlock".  Whoops!

Config changes were being written to the split tablets while the
tablet lock was held.  This had the potential to get into deadlock.
The solution elsewhere was to defer config changes until later, and
the same approach was taken here.  Config changes are queued up, but
the split completes without waiting for them to go through.

If the client calls sync(), it will block for the config change.  If
something crashes before the config change, the next server shouldn't
have any troubles recovering, though this assumption needs to be more
thoroughly tested.