Implemented a slightly smarter compaction algorithm
Instead of compacting the last 8 fragments from a Tablet, it selects a
contiguous window of up to 20 fragments from anywhere in the fragment
list.  The algorithm tries to choose the window with the smallest disk
footprint, on the assumption it will be faster to compact.

This should help somewhat when the compaction thread falls behind.
Lightweight sections in the middle can be replaced instead of
repeatedly compacting the data in the tail.  This still has the
problem of creating too many small files when logs are shared between
many Tablets.