Made Scanner reopening happen lazily
Scanner::reopen() now just releases the old scanner rather than trying
to open a new one.  Opening a new scan is deferred to the next call to
get().  This is nice, because if there's a problem reopening the
scanner, the exception is reported to the caller of get(), not some
internal maintenance thread that triggered the reopen().