Wrapped background thread loops with callOrDie()
They used to do some logging and then exit, leaving the rest of the
application to limp along.  We don't have a good way to recover from
errors, and the limping behavior is misleading and potentially
dangerous.  It's better to just shoot down the process with an
informative error message.

Also SharedLogger used to have some peculiar code for trying to get
the error condition from the worker thread back to one of the API
threads.  It's not clear that's particularly helpful.  Also, it didn't
seem to work very well.  The one time it executed, the process
starting calling pure virtuals shortly after and blew up.  Better to
keep it simple.