Updated clone forwarding to handle multiple cloned logs
The clone forwarding and duplicate detection code assumed that each
tablet can have only one log fragment at a time.  It turns out that's
not true.  A tablet keeps its handle to logs in the serialization
queue until they get replaced.  If updates are coming in quickly, it's
entirely possible for a tablet to have two logs while one is being
serialized to disk.  If serialization is going slowly, the tablet may
have as many logs as the serialization queue depth allows.

The new forwarding code allocates an independent forwarding chain for
each cloned log.  The duplicate detection code tracks specific
duplicate logs instead of a single flag on the last log.