Reworked compaction algo for non-adjacent fragments.  Sorta.
It turned out that the way I was expanding the compaction set gave me
subgraphs with holes in tablet fragment lists.  Fixing that seemed
hard, so I tried to make the algorithm work even when the compaction
subgraph has holes in it.

The algorithm tries to choose the longest adjacent subsequence in the
graph for each tablet, and then compacts only the parts of fragments
that are included in those subsequences.

The algo seems correct, but there's a bug in how I find the adjacent
subsequence.  FragDag's getParent() and getChild() are wrong.  Just
because a fragment is in the parent set of the query fragment and in
the fragment set of the query tablet does not mean that the fragment
is the query fragment's parent in the query tablet's fragment chain.
Oops.  Need to fix that.

Also, if we happen to choose a compaction set where no meaningful work
can be done, we'll throw an exception (and crash).  I'm not sure how
to guarantee that doesn't happen, or how to recover if it does.