Made compact() reorder/rebuild merge at each range boundary
This should work around the cycle problem in the fragment graph.  It
doesn't matter if there are cycles in the fragment graph since each
tablet has a clear prescribed fragment ordering.

WIP: this is placeholder code.  It assumes some cutoff functionality
for each input stream in the merge which isn't implemented yet.