Created kdiWatchdog script to keep kdiNetServer up
The script monitors a kdiNetServer PID file and make sure the process
is running an called kdiNetServer.  If the process goes away, the
watchdog will start a new one.  The watchdog will also periodically
scan from and load to a small table on server.  If these commands fail
or time out for long enough, the server is assumed to be wedged and
the watchdog will kill the process.

Added kdi control script support for starting and stopping the
watchdog (kdi watch and kdi unwatch).  Also added an argument to make
kdiNetServer write its own PID file.