Fixed incoherent synchronization of Tablet row range
We used to have a rows object that was supposed to be half thread safe
-- the upper bound was supposed to be constant and the lower bound
could change.  However, a lot of code completely lost sight of those
guidelines.  Some code always locked for rows access, some never
locked.  Also, even though the upper bound always had the same value,
it was still being rewritten, which is dangerous.

Removed rows member.  Replaced with const maxRow and mutable minRow.
Tried to make sure that anything that wants to work with minRow holds
the mutex.