Make NetTable retry for 15 minutes on connection failures
All mutations are now sync'd immediately.  If the connection fails
while applying or syncing a mutation, it will be reapplied if the
connection can be restored.

If the connection to the server drops, the client will attempt to
restore it once per minute for up to 15 minutes before giving up.