Removed TimeoutLocator, added ScannerLocator and TableLocator
TimeoutLocator was very confusing.  It had different behaviors for
scanners and tables, and it didn't work very well for scanners.
Replaced it with two specific locators.

TableLocator creates tables and never deletes them.  Rather than the
convoluted virtual Item logic it was using before, it now has a simple
map of objects, each marked by a state.  The loading code is more
understandable.  The timeout logic is all gone.

ScannerLocator uses a LruCache for Scanner objects instead of using a
timeout.  This gives us a fixed upper bound on the number of scanners
we have in the system and prevents crazy clients from crashing the
system by allocating too many scanners.

Also, scanners are now indentified by an increasing integer handle
instead of an Ice UUID.  This should save a couple of bytes on
transport overhead and a couple of cycles on servant lookup.