Added custom size and init functors to LruCache
The size of the LRU cache is now measured by a custom size functor
over value objects rather than object count.  The default
implementation is unit size per object.

When a new object is created in the pool, the cache will run an
initializer functor with the new value and key.  This allows the value
to be initialized from the key before measuring its size.  The default
initializer is a no-op.