Added safe decoding functions; added signed varint suite
The safe decoding functions take a max buffer length.  If the varint
decoding would exceed the buffer, the function returns zero.

The signed varint suite maps the signed integers to unsigned integers,
then uses the unsigned varint tools.  The mapping is as follows:

  0:0, -1:1, 1:2, -2:3, 2:4, -3:5, 3:6, ...

Non-negative numbers map to the even integers, and negative numbers
map to the odd integers, in magnitude order.