Updated kdiIntervals limited partitioning algo
Always produce the specified number of output intervals, assuming
there are enough input intervals.  In the case of cut intervals, the
cuts will be weighted by the number of input intervals contained in
the cut, though partitioning can get sloppy if significant chunks of
the input go into few cuts.