Made warp::less the default comparitor for Interval stuff
Also removed the optional comparitor from IntervalSet::contains.  The
ordering used must be consistent with the set's ordering, and the
easiest way to ensure that is to use the set's comparitor.  If we want
alternate type comparsion, the set's comparitor should support it
(which is a reason for changing the default to warp::less).