Split ConfigReader into SchemaReader and ConfigReader
Old ConfigReader was an async vector interface to TableSchemas and
TabletConfigs.  New ConfigReader and SchemaReader are synchronous,
one-at-a-time interfaces for TabletConfigs and TableSchemas,
respectively.

The async handling is done by the async load path.  The old vector
interface made things too complicated.  For a table-based config
reader, the bundling efficiency can probable be better accomplished by
read-ahead and caching in the implementation.

This is a little closer to working.