Hooked up faulty local Fragment GC; needs fixing
It's broken in that the same cached fragment may get multiple GC
wrappers on it, and the underlying file gets deleted when the first
wrapper is deleted, not the last.  A better solution is probably to
integrate better with the FileTracker stuff.  Soon.

Also, added a bit more work on Tablet unloading.