Reworked local fragment GC
The new local fragment GC is integrated into FileTracker (of
PendingFile fame).  The FileTracker is now a FragmentLoader
implementation, and fragments are tracked when they are loaded.  Added
a new exportFragment() method to the Fragment interface, which allows
Fragments to go untracked.

This reformulation of local GC should avoid problems with multiple
references and potential race conditions on newly created files (the
PendingFile counts as a reference).  As a minor bonus, if the
PendingFile never gets a loaded Fragment before being released, its
file will be cleaned up by the local GC.

Fragments are automatically exported on load.  Need to add automatic
export on Tablet unload/destruction.